requirements.txt
Django
djangorestframework-simplejwt
django-cors-headers
Pillow
psycopg2
djangorestframework
drf-yasg
gunicorn
